check_status.sh : add script to aid file based GTD
The check_status.sh script is meant to help maintain a GTD system based
on files.  The GTD system consists of a status file, status.txt, which
contains references to task files in subdirectories.  The subdirectories
represent the state of the task and any contents can be added to the
task files.  The check_status.sh script helps ensure that a) all of the
references in the status file have a corresponding task file on the
filesystem and b) all of the task files on the filesystem are present in
the status file.  The two checks can be executed individually via flags
otherwise both types of checks will be performed.