cryptar: similar to encedit but for encrypted tarballs
This is a work in progress. A bunch of code will be stolen/resude from
encedit. Perhaps the common code will be extracted to a single location
at some point. Cryptar will follow many of the same steps as encedit to
allow the contents of an encrypted tarball to modified in a relatively
safe manner.