sqlalchemy_sandbox.py: merged datastore and schema classes
I thought it would be necessary to have a Datastore object for lowest
level abstraction, a Schema object for a middle layer and then an App
subclass adding specific behaviour for the Relational DB Model. The
middle layer doesn't seem to be helpful to me right now so I removed it.