add a collection of bash scripts I use for work
When I work from home I use Cygwin on a Windows VM to log into the VPN
and then transfer the files that I'm working on back and forth. This way
I don't have to fight against latency so much. These scripts simplify my
workflow. Normally I would have used the Unison program but it has way
to many dependencies to get it to work easily on Windows.