copy: make output flags a config file option
A few of the different VMs that I have access to apparently have
different versions of rsync with different levels of support for the
more granular output options that were added in newer versions. This
lets me use the backwards compatible "-vv" on older rsync versions and
newer incantations like "--info=progress2" with newer versions.